- Resource IDs are no longer generated by the server - Resource IDs are now only unique in combination with a user id - Use names to get data from rows

+++ src/database/member.rs
@@ -6,8 +6,6 @@ use sqlx::mysql::MySqlRow;
 use sqlx::{query, Row};
 
 pub async fn get_members(pool: &DatabasePool, user_id: UserId) -> DatabaseResult<Vec<Member>> {
-    const ID_IDX: usize = 10;
-
     let members = query("SELECT UserId, Name, Pronouns, AvatarUrl, Description, Color, CreatedAt, UpdatedAt, Archived, Custom, ID FROM Member WHERE UserId = ?")
         .bind(user_id)
         .fetch_all(pool.as_ref())
@@ -22,7 +20,7 @@ pub async fn get_members(pool: &DatabasePool, user_id: UserId) -> DatabaseResult
         let sql = format!("SELECT MemberId, FolderId FROM MemberFolder WHERE MemberId IN ({placeholders})");
         let mut query = query(sql.as_str());
         for member in &members {
-            let id: MemberId = member.get(ID_IDX);
+            let id: MemberId = member.get("ID");
             query = query.bind(id);
         }
         query
@@ -31,7 +29,7 @@ pub async fn get_members(pool: &DatabasePool, user_id: UserId) -> DatabaseResult
     };
 
     Ok(members.into_iter().map(|row| {
-        let id = row.get(ID_IDX);
+        let id = row.get("ID");
         let folders = folders.iter()
             .filter(|row| {
                 let member_id: MemberId = row.get(0);
@@ -43,9 +41,10 @@ pub async fn get_members(pool: &DatabasePool, user_id: UserId) -> DatabaseResult
     }).collect())
 }
 
-pub async fn get_member_by_id(pool: &DatabasePool, member_id: MemberId) -> DatabaseResult<Option<Member>> {
-    let res = query("SELECT UserId, Name, Pronouns, AvatarUrl, Description, Color, CreatedAt, UpdatedAt, Archived, Custom FROM Member WHERE ID = ?")
+pub async fn get_member_by_id(pool: &DatabasePool, member_id: MemberId, user_id: UserId) -> DatabaseResult<Option<Member>> {
+    let res = query("SELECT UserId, Name, Pronouns, AvatarUrl, Description, Color, CreatedAt, UpdatedAt, Archived, Custom FROM Member WHERE ID = ? AND UserId = ?")
         .bind(member_id)
+        .bind(user_id)
         .fetch_optional(pool.as_ref())
         .await?;
 
@@ -54,25 +53,18 @@ pub async fn get_member_by_id(pool: &DatabasePool, member_id: MemberId) -> Datab
     }
     let res = res.unwrap();
 
-    let folders = query("SELECT FolderId FROM MemberFolder WHERE MemberId = ?")
+    let folders = query("SELECT FolderId FROM MemberFolder WHERE MemberId = ? AND UserId = ?")
         .bind(member_id)
+        .bind(user_id)
         .fetch_all(pool.as_ref())
         .await?;
 
     Ok(Some(member(res, member_id, folders.into_iter().map(|row| row.get(0)).collect())))
 }
 
-pub async fn get_member_user_id(pool: &DatabasePool, member_id: MemberId) -> DatabaseResult<UserId> {
-    let user_id = query("SELECT UserId FROM Member WHERE ID = ?")
-        .bind(member_id)
-        .fetch_one(pool.as_ref())
-        .await?;
-
-    Ok(user_id.get(0))
-}
-
-pub async fn create_member(pool: &DatabasePool, user_id: UserId, member: &Member) -> DatabaseResult<MemberId> {
-    let res = query("INSERT INTO Member (UserId, Name, Pronouns, AvatarUrl, Description, Color, Custom) VALUES (?, ?, ?, ?, ?, ?, ?) RETURNING ID")
+pub async fn create_member(pool: &DatabasePool, user_id: UserId, member: &Member) -> DatabaseResult<()> {
+    query("INSERT INTO Member (ID, UserId, Name, Pronouns, AvatarUrl, Description, Color, Custom) VALUES (?, ?, ?, ?, ?, ?, ?, ?)")
+        .bind(member.id)
         .bind(user_id)
         .bind(&member.name)
         .bind(&member.pronouns)
@@ -80,10 +72,10 @@ pub async fn create_member(pool: &DatabasePool, user_id: UserId, member: &Member
         .bind(&member.description)
         .bind(member.color)
         .bind(member.custom)
-        .fetch_one(pool.as_ref())
+        .execute(pool.as_ref())
         .await?;
 
-    Ok(res.get(0))
+    Ok(())
 }
 
 pub async fn delete_member(pool: &DatabasePool, member_id: MemberId, user_id: UserId) -> DatabaseResult<()> {
@@ -111,16 +103,18 @@ pub async fn edit_member(pool: &DatabasePool, member: &Member) -> DatabaseResult
     Ok(())
 }
 
-pub async fn edit_member_folders(pool: &DatabasePool, member_id: MemberId, folder_ids: &[FolderId]) -> DatabaseResult<()> {
+pub async fn edit_member_folders(pool: &DatabasePool, member_id: MemberId, user_id: UserId, folder_ids: &[FolderId]) -> DatabaseResult<()> {
     let mut tx = pool.begin().await?;
 
-    query("DELETE FROM MemberFolder WHERE MemberId = ?")
+    query("DELETE FROM MemberFolder WHERE MemberId = ? AND UserId = ?")
         .bind(member_id)
+        .bind(user_id)
         .execute(&mut *tx)
         .await?;
 
     for folder_id in folder_ids {
-        query("INSERT INTO MemberFolder (MemberId, FolderId) VALUES (?, ?)")
+        query("INSERT INTO MemberFolder (UserId, MemberId, FolderId) VALUES (?, ?, ?)")
+            .bind(user_id)
             .bind(member_id)
             .bind(folder_id)
             .execute(&mut *tx)
@@ -132,16 +126,16 @@ pub async fn edit_member_folders(pool: &DatabasePool, member_id: MemberId, folde
 }
 
 fn member(row: MySqlRow, id: MemberId, folders: Vec<FolderId>) -> Member {
-    let user_id = row.get(0);
-    let name = row.get(1);
-    let pronouns = row.get(2);
-    let avatar = row.get(3);
-    let description = row.get(4);
-    let color = row.get(5);
-    let created_at = row.get(6);
-    let updated_at = row.get(7);
-    let archived = row.get(8);
-    let custom = row.get(9);
+    let user_id = row.get("UserId");
+    let name = row.get("Name");
+    let pronouns = row.get("Pronouns");
+    let avatar = row.get("AvatarUrl");
+    let description = row.get("Description");
+    let color = row.get("Color");
+    let created_at = row.get("CreatedAt");
+    let updated_at = row.get("UpdatedAt");
+    let archived = row.get("Archived");
+    let custom = row.get("Custom");
 
     Member {
         id,
